Robotmk: Separate check plugin for suite status
CMK-14987
