+++ checkman/robotmk_suite
@@ -1,0 +1,10 @@
+title: Robotmk Suite
+agents: robotmk
+catalog: robotframework
+license: GPLv2
+distribution: check_mk
+description:
+ Provides information about the status of a suite.
+
+discovery:
+ One service is per suite is created.
